Prevent order deletion if already assigned

+++ src/main/java/seedu/address/logic/commands/order/DeleteCommand.java
@@ -42,6 +42,10 @@ public class DeleteCommand extends OrderCommand {
         }
 
         Order orderToDelete = lastShownList.get(targetIndex.getZeroBased());
+        if (orderToDelete.isAlreadyAssignedDeliveryman()) {
+            throw new CommandException(Messages.MESSAGE_ORDER_HAS_DELIVERYMAN_CANNOT_DELETE);
+        }
+
         model.deleteOrder(orderToDelete);
         model.commitOrderBook();
         return new CommandResult(String.format(MESSAGE_DELETE_ORDER_SUCCESS, orderToDelete));
